+++ src/extension/httpOutputProvider/rfc7230HttpOutputProvider.ts
@@ -1,6 +1,6 @@
 import type * as Httpyac from 'httpyac';
 import * as vscode from 'vscode';
-import { AppConfig } from '../config';
+import { AppConfig, Rfc7230Output } from '../config';
 import { HttpOutputProvider, HttpOutputResult, HttpOutputPriority } from '../extensionApi';
 
 export class Rfc7230HttpOutputProvider implements HttpOutputProvider {
@@ -9,6 +9,29 @@ export class Rfc7230HttpOutputProvider implements HttpOutputProvider {
   constructor(readonly config: AppConfig, readonly httpyac: typeof Httpyac) {}
 
   getResponseOutputResult(response: Httpyac.HttpResponse): HttpOutputResult | false {
+    const httpResponse: Httpyac.HttpResponse = this.prepareHttpResponse(response);
+
+    if (this.config.outputRfc7230 === Rfc7230Output.response) {
+      delete httpResponse.request;
+    } else if (this.config.outputRfc7230 === Rfc7230Output.request_header_and_response) {
+      delete httpResponse.request?.body;
+    } else if (this.config.outputRfc7230 === Rfc7230Output.only_header) {
+      delete httpResponse.request?.body;
+      delete httpResponse.body;
+      delete httpResponse.prettyPrintBody;
+    } else if (this.config.outputRfc7230 === Rfc7230Output.only_response_header) {
+      delete httpResponse.request;
+      delete httpResponse.body;
+      delete httpResponse.prettyPrintBody;
+    }
+
+    return {
+      outputItems: vscode.NotebookCellOutputItem.json(httpResponse, 'message/http'),
+      priority: HttpOutputPriority.Default,
+    };
+  }
+
+  private prepareHttpResponse(response: Httpyac.HttpResponse) {
     const httpResponse: Httpyac.HttpResponse = {
       ...response,
       request: response.request && {
@@ -23,10 +46,8 @@ export class Rfc7230HttpOutputProvider implements HttpOutputProvider {
         'base64'
       )}`;
     }
-
-    return {
-      outputItems: vscode.NotebookCellOutputItem.json(httpResponse, 'message/http'),
-      priority: HttpOutputPriority.Default,
-    };
+    delete httpResponse.rawBody;
+    delete httpResponse.parsedBody;
+    return httpResponse;
   }
 }